Background script
Addon now notices when the tab URL changes through a background script. Because of that it only does things when it notices the "watch" in the URL. It also searches for the chat-div and changes the calculation based on if it finds it or not.

Also removed the style.css-file, since it was not needed anymore.