Less scrolling to the top & fullscreen support
The page gets scrolled to the top much less (for example, it no longer scrolls to the top after the video ends), but sometimes is not correctly at the top. The video now most of the time works with full screen mode. Sometimes the aria-label of the videoParent changes slower than normal so the size is incorrect, but it usually gets fixed by going out of and back into full screen mode. Also removed a console.log from the background script and an unnecessary variable inWatch.