Youtube changed again
Yoututube changed the element that you need to increase the size of, so the addon now searches and changes the corect element.
Starting from previous searches, Youtube did not handle videos with chat well (could have been partly due to other addons). That has been dixed by waiting for a chat  element and calling a resize event when it is found. Might not be needed in the future.